REFACTOR: rearrange resource package
