+++ execute/src/main/kotlin/uk/dioxic/mgenerate/execute/resources/Resources.kt
@@ -1,9 +1,15 @@
 package uk.dioxic.mgenerate.execute.resources
 
-import arrow.fx.coroutines.ResourceScope
-import com.mongodb.MongoClientSettings
 import com.mongodb.client.MongoClient
-import com.mongodb.client.MongoClients
+import com.mongodb.client.MongoCollection
+import com.mongodb.client.MongoDatabase
 
-suspend fun ResourceScope.mongoClient(settings: MongoClientSettings): MongoClient =
-    install({ MongoClients.create(settings) }) { client, _ -> client.close() }
+sealed interface Resource
+
+class MongoResource(val client: MongoClient) : Resource {
+    fun getDatabase(database: String): MongoDatabase =
+        client.getDatabase(database)
+
+    inline fun <reified TDocument> getCollection(database: String, collections: String): MongoCollection<TDocument> =
+        client.getDatabase(database).getCollection(collections, TDocument::class.java)
+}
