+++ execute/src/test/kotlin/uk/dioxic/mgenerate/execute/ResourceRegistryTest.kt
@@ -1,0 +1,9 @@
+package uk.dioxic.mgenerate.execute
+
+import io.kotest.core.spec.style.FunSpec
+
+class ResourceRegistryTest: FunSpec({
+
+    test("")
+
+})
